requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
torch==2.0.1
torchvision==0.15.2
numpy==1.25.2
pillow==10.0.0
scikit-learn==1.3.0
tqdm==4.66.1
